Upgrade to easyfft v0.3
Unfortunately this requires an extra Default trait bound which results
in requiring a major version change for fftconvolve.

Practically it's not an issue since all elements that you'd want to
convolve would implement Default, but it's a bit annoying.

Default is required for initializing the scratch-buffer, which is
strictly not necessary but requires changes to `rustfft`. It's currently
pending this issue in rustfft:
https://github.com/ejmahler/RustFFT/issues/105